Move progress file
