+++ src/pytorch_lightning/loops/epoch/prediction_epoch_loop.py
@@ -5,8 +5,8 @@ import torch
 
 from lightning_fabric.utilities import move_data_to_device
 from pytorch_lightning.loops.loop import _Loop
+from pytorch_lightning.loops.progress import Progress
 from pytorch_lightning.overrides.distributed import IndexBatchSamplerWrapper
-from pytorch_lightning.trainer.progress import Progress
 from pytorch_lightning.utilities.rank_zero import WarningCache
 
 warning_cache = WarningCache()
