+++ tests/tests_pytorch/loops/test_loops.py
@@ -26,7 +26,7 @@ from pytorch_lightning import LightningModule, Trainer
 from pytorch_lightning.callbacks import Callback, ModelCheckpoint
 from pytorch_lightning.demos.boring_classes import BoringModel, RandomDataset
 from pytorch_lightning.loops import _Loop
-from pytorch_lightning.trainer.progress import BaseProgress
+from pytorch_lightning.loops.progress import BaseProgress
 from tests_pytorch.helpers.runif import RunIf
