+++ tests/tests_pytorch/loops/test_progress.py
@@ -15,7 +15,7 @@ from copy import deepcopy
 
 import pytest
 
-from pytorch_lightning.trainer.progress import (
+from pytorch_lightning.loops.progress import (
     BaseProgress,
     OptimizerProgress,
     ProcessedTracker,
